pyproject.toml
[project]
name = "commerce-ai-platform"
version = "0.1.0"
description = "Production-grade FastAPI backend skeleton for Commerce AI Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.4.0",
    "openai-whisper>=20231117",
    "anthropic>=0.39.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
